ref: major, incomplete overhaul of should_exec
This is part 1 of 2. In the end, the actual "logic" or algorithm of
`_should_execute` will NOT have changed, but both the code and
warnings/notifications will have changed substantially. This is to make
it more explicit and more clear what is happening in each scenario.

The second part of this change is small, and will include a new CLI
argument that has explicit options. I'll write it up in the docstring
for `main`.