ref: nb_json_output_dir vs pics output
Per notebook, there are really "two" output folders: one that has the
"JSON output" of the run notebook (and possibly the "standalone HTML"),
and a second one inside that dir named "output_nb_<notebook>" that
stores images that are output from the execution.

Eventually, *all* of the output from a notebook should be contained in
its "output_nb_<notebook>" dir, including the "JSON output". For now, we
will deal with them separately, since they are both valid "output directories".