[WIP]: REF: Refactor of new dev build process code (#122)
* squashed refactor commits to rebase

* [no ci] fix: dev dir was not creating subdirs

* [no ci] harmonize Makefile, remove test hash.json

* ref: move MPI deps into env file, refactor make

* [no ci]upgrade nbs to work with latest hnn version

* [no ci]ref: move gen pages into its own module

* ref: move version hash to module, rename

* [no ci] ref: several renames

* ref: improve naming of skips, separate load func

* ref: where possible, replace notebook with nb

This has not been changed in documentation, or user-facing CLI
arguments, or HTML. This change has been done in internal python
variables, functions, and filenames, since I'm tired of typing out
"notebook"!

* fix: exclude no-src cells from exec test

* ref: improve clarity of exec-proc ctrl flow

* ref: ruff

* [no ci] ref: add initial notes for exec-algo ref

* ref: minor note comments

* ref: get_page_paths renamed and ref with Pathlib

* [no ci] ref: s/os/pathlib/g in build, add arg

* ref: add passthrough of nb_skip_path and minor

* ref: minor

* ref: simplify _nb_has_json_output and output_dir

* ref: major, incomplete overhaul of should_exec

This is part 1 of 2. In the end, the actual "logic" or algorithm of
`_should_execute` will NOT have changed, but both the code and
warnings/notifications will have changed substantially. This is to make
it more explicit and more clear what is happening in each scenario.

The second part of this change is small, and will include a new CLI
argument that has explicit options. I'll write it up in the docstring
for `main`.

* ref: major overhaul of CLI args, new exec_filter

* fix: make deploy GHA use new conda recipe and name

* fix: accomodate new check_nb_version location

* doc: use Claude for docs of _process_nb and subf's

Also includes a ruff format and a minor private function name change.

* ref: use textwrap.dedent to prettify nb warnings

* ref: nb_json_output_dir vs pics output

Per notebook, there are really "two" output folders: one that has the
"JSON output" of the run notebook (and possibly the "standalone HTML"),
and a second one inside that dir named "output_nb_<notebook>" that
stores images that are output from the execution.

Eventually, *all* of the output from a notebook should be contained in
its "output_nb_<notebook>" dir, including the "JSON output". For now, we
will deal with them separately, since they are both valid "output directories".

* ref: replace nb out img save with cleaner pathlib

This also makes it clearer that what was formerly the "input dir" for
the image-saving code is actually the "JSON output dir"

* doc: use claude for docs of _extract_html_from_nb

* ref: clean with pathlib, and gen doc w/ Claude

* maint: remove unused func

* ref: del and replace standalone with pathlib

* ref: various renames and more docs

* ref: for nb exec f, add docs, use pathlib for walk

* ref: add return content/dev path from execute func

* ref: get_mds to gen_page_html, and consume content

* maint: add dev html removal to make clean

* fix: add capture and doc of case of prior_version

* ref: more path-separation ref and docs

* ref: separate and ref index generation

* ref: revert index relocation, start reintegrating

* forgot to add new file

* reenable full build process

* ref: cleanup comments and turn todos into Issues

---------

Co-authored-by: dylansdaniels <dylansdaniels@gmail.com>